Split models package for easier maintenance
